python-backend/requirements.txt
flask
requests
gunicorn
python-dotenv
flask-cors
